LayoutTests/media/track/track-webvtt-tc006-cue-identifiers-expected.txt

Tests that any text other than "-->" is recognized as optional cue identifier.


*** Testing text track 0
EXPECTED (cues.length == '4') OK
EXPECTED (cues[i].id == 'random_id') OK
EXPECTED (cues[i].startTime == '0') OK
EXPECTED (cues[i].endTime == '30.5') OK
EXPECTED (cues[i].getCueAsSource() == 'Bear is Coming!!!!!') OK
EXPECTED (cues[i].id == 'another random identifier') OK
EXPECTED (cues[i].startTime == '31') OK
EXPECTED (cues[i].endTime == '60.5') OK
EXPECTED (cues[i].getCueAsSource() == 'I said Bear is coming!!!!') OK
EXPECTED (cues[i].id == 'identifier--too') OK
EXPECTED (cues[i].startTime == '61') OK
EXPECTED (cues[i].endTime == '120.5') OK
EXPECTED (cues[i].getCueAsSource() == 'I said Bear is coming now!!!!') OK
EXPECTED (cues[i].id == 'identifier--too') OK
EXPECTED (cues[i].startTime == '121') OK
EXPECTED (cues[i].endTime == '180.5') OK
EXPECTED (cues[i].getCueAsSource() == 'Duplicate identifier') OK

*** Testing text track 1
EXPECTED (cues.length == '0') OK
END OF TEST
